# dist/utils/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NFS API Documentation",
        "version": "1.0.0",
        "description": "Documentation for the NFS Backend API"
    },
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "Development Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/admin/users/{id}": {
            "patch": {
                "tags": ["Admin - Users"],
                "summary": "Update user status (Activate/Block)",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "Status updated" } }
            }
        },
        "/api/admin/users/{id}/credit": {
            "post": {
                "tags": ["Admin - Users"],
                "summary": "Initiate credit on user account",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "Credit requested" } }
            }
        },
...
        "/api/admin/transactions/{txId}/validate": {
            "put": {
                "tags": ["Admin - Transactions"],
                "summary": "Validate transaction (COMEX)",
                "parameters": [
                    { "name": "txId", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "Transaction validated" } }
            }
        },
        "/api/admin/transactions/{txId}/reject": {
            "put": {
                "tags": ["Admin - Transactions"],
                "summary": "Reject transaction (COMEX)",
                "parameters": [
                    { "name": "txId", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "Transaction rejected" } }
            }
        },
...
        "/api/admin/groups/{id}": {
            "put": {
                "tags": ["Admin - Security"],
                "summary": "Update a group",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "Group updated" } }
            }
        },
        "/api/admin/users/{id}/groups": {
            "put": {
                "tags": ["Admin - Security"],
                "summary": "Assign groups to user",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "Groups assigned" } }
            }
        },
...
        "/api/admin/loans/{id}": {
            "patch": {
                "tags": ["Admin - Products"],
                "summary": "Update loan status",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": { "200": { "description": "Loan updated" } }
            }
        },
...
    }
}